+++ ateliersoude/event/tests/test_condition.py
@@ -63,8 +63,8 @@ def test_condition_delete(client, user_log, condition_factory):
     assert Condition.objects.count() == 0
     assert response.status_code == 302
     assert response["Location"] == reverse(
-        "user:organization_detail",
-        args=[condition.organization.pk, condition.organization.slug],
+        "organization_detail",
+        args=[condition.organization.slug],
     )
 
 
@@ -115,7 +115,7 @@ def test_condition_create(client, user_log, condition_data, organization):
     assert len(conditions) == 1
     orga = conditions[0].organization
     assert response["Location"] == reverse(
-        "user:organization_detail", args=[orga.pk, orga.slug]
+        "organization_detail", args=[orga.slug]
     )
 
 
@@ -181,7 +181,7 @@ def test_condition_update(
     orga = conditions[0].organization
     assert conditions[0].name == "cond_name2"
     assert response["Location"] == reverse(
-        "user:organization_detail", args=[orga.pk, orga.slug]
+        "organization_detail", args=[orga.slug]
     )
