+++ ateliersoude/templates/menu.html
@@ -70,7 +70,7 @@
           {% if user.admin_organizations.all %}
           <div class="dropdown-divider"></div>
           {% for organization in user.admin_organizations.all %}
-            <a class="dropdown-item" href="{% url "user:organization_detail" organization.pk organization.slug %}">
+            <a class="dropdown-item" href="{% url "organization_detail" organization.slug %}">
               Gérer {{ organization.name }}
             </a>
           {% endfor %}
