+++ ateliersoude/user/templates/user/organization/confirmation_delete.html
@@ -2,7 +2,7 @@
 
 {% block content %}
   {% url "user:organization_list" as organization_list_url %}
-  {% url "user:organization_detail" organization.pk organization.slug as organization_detail_url %}
+  {% url "organization_detail" organization.pk organization.slug as organization_detail_url %}
   {% include "breadcrumb.html" with current="Suppression" second_parent_url=organization_list_url second_parent_text="Organisations" first_parent_url=organization_detail_url first_parent_text=organization %}
 
   <div class="jumbotron jumbotron-fluid">
