+++ ateliersoude/user/templates/user/organization/organization_form.html
@@ -5,7 +5,7 @@
 {% block content %}
 {% url "user:organization_list" as organization_list_url %}
 {% if organization %}
-  {% url "user:organization_detail" organization.pk organization.slug as organization_detail_url %}
+  {% url "organization_detail" organization.slug as organization_detail_url %}
   {% include "breadcrumb.html" with current="Mise à jour" second_parent_url=organization_list_url second_parent_text="Organisations" first_parent_url=organization_detail_url first_parent_text=organization %}
 {% else %}
   {% include "breadcrumb.html" with current="Création" first_parent_url=organization_list_url first_parent_text="Organisations" %}
